changed pool interface around thus:
- the inline functions now get defined in the user's header, rather than in their source file
- the pool singleton is globally visible, rather than static, and is declared extern by the header

this means that functions which call the inlined pool functions can themselves be inlined, as
they no longer need to reference the static singleton object.

this, in turn, means that pooled scalars, which are a very light wrapper around anonymous scalars,
potentially don't require two function calls for every operation if the wrappers are inlined.